requirements.txt
# FastAPI Foundation
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Logfire Observability
logfire[fastapi]>=0.28.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0  # For PostgreSQL async support

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For async testing
pytest-mock>=3.11.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pandera>=0.17.0  # Data validation
pyarrow>=14.0.0  # For Parquet files

# Genetic Algorithm (Darwin)
pygad>=3.2.0  # Genetic algorithm framework
deap>=1.4.1  # Alternative GA library

# Panel Dashboard
panel>=1.3.0
bokeh>=3.3.0
param>=2.0.0
hvplot>=0.9.0

# Stream Processing Simulation
aiokafka>=0.10.0  # Kafka client for Kinesis simulation
redis>=5.0.0  # For pub/sub and caching

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Development Tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.3.0

# API Documentation
email-validator>=2.0.0

# Task Scheduling (for Airflow simulation)
apscheduler>=3.10.0

# Data Quality
great-expectations>=0.18.0

# Infrastructure as Code
pulumi>=3.0.0
pulumi-aws>=6.0.0

# Utilities
httpx>=0.25.0
orjson>=3.9.0  # Fast JSON
python-dateutil>=2.8.0
pytz>=2023.3
tenacity>=8.2.0  # Retry logic

# Monitoring & Metrics
prometheus-client>=0.19.0

# TypeScript/Frontend Build Support
# (These are typically installed separately but documenting for reference)
# nodejs>=18.0.0
# npm>=9.0.0
